+++ dd-java-agent/instrumentation/spring-webflux/src/main/java/datadog/trace/instrumentation/springwebflux/server/DispatcherHandlerInstrumentation.java
@@ -1,12 +1,9 @@
-package datadog.trace.instrumentation.springwebflux;
+package datadog.trace.instrumentation.springwebflux.server;
 
-import static datadog.trace.agent.tooling.ByteBuddyElementMatchers.safeHasSuperType;
 import static java.util.Collections.singletonMap;
-import static net.bytebuddy.matcher.ElementMatchers.isAbstract;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.not;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
@@ -18,20 +15,11 @@ import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
 @AutoService(Instrumenter.class)
-public final class RouterFunctionInstrumentation extends AbstractWebfluxInstrumentation {
-
-  public RouterFunctionInstrumentation() {
-    super("spring-webflux-functional");
-  }
+public final class DispatcherHandlerInstrumentation extends AbstractWebfluxInstrumentation {
 
   @Override
   public ElementMatcher<TypeDescription> typeMatcher() {
-    return not(isAbstract())
-        .and(
-            safeHasSuperType(
-                // TODO: this doesn't handle nested routes (DefaultNestedRouterFunction)
-                named(
-                    "org.springframework.web.reactive.function.server.RouterFunctions$DefaultRouterFunction")));
+    return named("org.springframework.web.reactive.DispatcherHandler");
   }
 
   @Override
@@ -39,12 +27,10 @@ public final class RouterFunctionInstrumentation extends AbstractWebfluxInstrume
     return singletonMap(
         isMethod()
             .and(isPublic())
-            .and(named("route"))
-            .and(
-                takesArgument(
-                    0, named("org.springframework.web.reactive.function.server.ServerRequest")))
+            .and(named("handle"))
+            .and(takesArgument(0, named("org.springframework.web.server.ServerWebExchange")))
             .and(takesArguments(1)),
         // Cannot reference class directly here because it would lead to class load failure on Java7
-        packageName + ".RouterFunctionAdvice");
+        packageName + ".DispatcherHandlerAdvice");
   }
 }
