+++ dd-java-agent/instrumentation/spring-webflux/src/main/java/datadog/trace/instrumentation/springwebflux/server/RouterFunctionInstrumentation.java
@@ -1,9 +1,12 @@
-package datadog.trace.instrumentation.springwebflux;
+package datadog.trace.instrumentation.springwebflux.server;
 
+import static datadog.trace.agent.tooling.ByteBuddyElementMatchers.safeHasSuperType;
 import static java.util.Collections.singletonMap;
+import static net.bytebuddy.matcher.ElementMatchers.isAbstract;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.not;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
@@ -15,11 +18,20 @@ import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
 @AutoService(Instrumenter.class)
-public final class DispatcherHandlerInstrumentation extends AbstractWebfluxInstrumentation {
+public final class RouterFunctionInstrumentation extends AbstractWebfluxInstrumentation {
+
+  public RouterFunctionInstrumentation() {
+    super("spring-webflux-functional");
+  }
 
   @Override
   public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("org.springframework.web.reactive.DispatcherHandler");
+    return not(isAbstract())
+        .and(
+            safeHasSuperType(
+                // TODO: this doesn't handle nested routes (DefaultNestedRouterFunction)
+                named(
+                    "org.springframework.web.reactive.function.server.RouterFunctions$DefaultRouterFunction")));
   }
 
   @Override
@@ -27,10 +39,12 @@ public final class DispatcherHandlerInstrumentation extends AbstractWebfluxInstr
     return singletonMap(
         isMethod()
             .and(isPublic())
-            .and(named("handle"))
-            .and(takesArgument(0, named("org.springframework.web.server.ServerWebExchange")))
+            .and(named("route"))
+            .and(
+                takesArgument(
+                    0, named("org.springframework.web.reactive.function.server.ServerRequest")))
             .and(takesArguments(1)),
         // Cannot reference class directly here because it would lead to class load failure on Java7
-        packageName + ".DispatcherHandlerAdvice");
+        packageName + ".RouterFunctionAdvice");
   }
 }
