+++ dd-java-agent/instrumentation/spring-webflux/src/main/java8/datadog/trace/instrumentation/springwebflux/server/RouterFunctionAdvice.java
@@ -1,4 +1,4 @@
-package datadog.trace.instrumentation.springwebflux;
+package datadog.trace.instrumentation.springwebflux.server;
 
 import net.bytebuddy.asm.Advice;
 import org.springframework.web.reactive.function.server.HandlerFunction;
@@ -14,10 +14,10 @@ public class RouterFunctionAdvice {
 
   @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
   public static void methodExit(
-    @Advice.This final RouterFunction thiz,
-    @Advice.Argument(0) final ServerRequest serverRequest,
-    @Advice.Return(readOnly = false) Mono<HandlerFunction<?>> result,
-    @Advice.Thrown final Throwable throwable) {
+      @Advice.This final RouterFunction thiz,
+      @Advice.Argument(0) final ServerRequest serverRequest,
+      @Advice.Return(readOnly = false) Mono<HandlerFunction<?>> result,
+      @Advice.Thrown final Throwable throwable) {
     if (throwable == null) {
       result = result.doOnSuccessOrError(new RouteOnSuccessOrError(thiz, serverRequest));
     } else {
