+++ dd-java-agent/instrumentation/spring-webflux/src/test/groovy/SingleThreadedSpringWebfluxTest.groovy
@@ -1,4 +1,4 @@
-import dd.trace.instrumentation.springwebflux.SpringWebFluxTestApplication
+import dd.trace.instrumentation.springwebflux.server.SpringWebFluxTestApplication
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.boot.test.context.TestConfiguration
 import org.springframework.boot.web.embedded.netty.NettyReactiveWebServerFactory
