+++ dd-java-agent/instrumentation/spring-webflux/src/test/groovy/dd/trace/instrumentation/springwebflux/server/FooModel.groovy
@@ -1,4 +1,4 @@
-package dd.trace.instrumentation.springwebflux
+package dd.trace.instrumentation.springwebflux.server
 
 class FooModel {
   public long id
