+++ dd-java-agent/instrumentation/spring-webflux/src/test/groovy/dd/trace/instrumentation/springwebflux/server/TestController.groovy
@@ -1,4 +1,4 @@
-package dd.trace.instrumentation.springwebflux
+package dd.trace.instrumentation.springwebflux.server
 
 import datadog.trace.api.Trace
 import org.springframework.web.bind.annotation.GetMapping
