+++ dd-java-agent/instrumentation/spring-webflux/src/test/java/dd/trace/instrumentation/springwebflux/server/RedirectComponent.java
@@ -1,4 +1,4 @@
-package dd.trace.instrumentation.springwebflux;
+package dd.trace.instrumentation.springwebflux.server;
 
 import static org.springframework.web.reactive.function.server.RequestPredicates.GET;
 import static org.springframework.web.reactive.function.server.RouterFunctions.route;
